+++ lib/Elastica/Node.php
@@ -142,20 +142,4 @@ class Node
         $this->_stats = null;
         $this->_info = null;
     }
-
-    /**
-     * Shuts this node down.
-     *
-     * @param string $delay OPTIONAL Delay after which node is shut down (default = 1s)
-     *
-     * @return \Elastica\Response
-     *
-     * @link https://www.elastic.co/guide/en/elasticsearch/reference/current/cluster-nodes-shutdown.html
-     */
-    public function shutdown($delay = '1s')
-    {
-        $path = '_cluster/nodes/'.$this->getId().'/_shutdown?delay='.$delay;
-
-        return $this->_client->request($path, Request::POST);
-    }
 }